int64

+++ day7_test.go
@@ -8,7 +8,7 @@ import (
 func TestMaxThrust(t *testing.T) {
 	for _, scen := range []struct {
 		program string
-		max     int
+		max     int64
 	}{
 		{program: "3,15,3,16,1002,16,10,16,1,16,15,15,4,15,99,0,0", max: 43210},
 		{program: "3,23,3,24,1002,24,10,24,1002,23,-1,23,101,5,23,23,1,24,23,23,4,23,99,0,0", max: 54321},
@@ -24,8 +24,8 @@ func TestMaxThrust(t *testing.T) {
 func TestMaxThrustWithFeedback(t *testing.T) {
 	for _, s := range []struct {
 		program string
-		max     int
-	} {
+		max     int64
+	}{
 		{program: "3,26,1001,26,-4,26,3,27,1002,27,2,27,1,27,26,27,4,27,1001,28,-1,28,1005,28,6,99,0,0,5", max: 139629729},
 	} {
 		thrust, err := maxThrustersWithFeedback(s.program)
